Add files via upload
This Zip file contains a code that includes the implementation of a sleep mode in the ESP8266 controller with a timer wakeup call. There are numerous sleep modes and in this, we have used deep sleep mode in which the CPUs, most of the RAM, and all digital peripherals are disabled except for the ULP Coprocessor, RTC Controller, RTC Peripherals, and RTC fast and slow memory. 

This mode is generally used where the CPU needs to be woken up by an external event, a timer, or a combination of these events while maintaining minimal power consumption. The real-time clock controller (RTC) of ESP8266 has a built-in timer that we can use to wake up ESP32 from a deep sleep after a predefined time has expired.